+++ node_editor/examples/empty.rs
@@ -35,19 +35,19 @@ impl ConstantEditor<primitives::IOValue> for MyConstantEditor {
             &mut IOValue::Float(ref mut float) => {
                 let mut out = *float as f32;
                 ui.input_float(im_str!("Float value"), &mut out).build();
-                *float = out as f64;
+                *float = out as f32;
             }
             &mut IOValue::Float2(ref mut floats) => {
                 let [a, b] = *floats;
                 let mut out = [a as f32, b as f32];
                 ui.input_float2(im_str!("2 floats value"), &mut out).build();
-                *floats = [out[0] as f64, out[1] as f64];
+                *floats = [out[0] as f32, out[1] as f32];
             }
             &mut IOValue::Float3(ref mut floats) => {
                 let [a, b, c] = *floats;
                 let mut out = [a as f32, b as f32, c as f32];
                 ui.input_float3(im_str!("3 floats value"), &mut out).build();
-                *floats = [out[0] as f64, out[1] as f64, out[2] as f64];
+                *floats = [out[0] as f32, out[1] as f32, out[2] as f32];
             }
             _ => (),
         }
